Add files via upload
Home Page is named by index. There are three Icons on the top right of header. When we log in or Sign up for order food. We have to click on user icon. Log-in and Sign-up option are here. If we click on Sign-up button. We find registration form to register our name . If we are sign-in already. We only log in. 
In "Our Food" Section, There are some fast food. If we need to order food . We click on "Order Now" button to order food.